Some clean up on room details
