+++ appnav/src/main/kotlin/io/element/android/appnav/RoomFlowNode.kt
@@ -18,8 +18,8 @@ package io.element.android.appnav
 
 import android.os.Parcelable
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.ui.Modifier
+import androidx.lifecycle.lifecycleScope
 import com.bumble.appyx.core.composable.Children
 import com.bumble.appyx.core.lifecycle.subscribe
 import com.bumble.appyx.core.modality.BuildContext
@@ -45,6 +45,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
+import kotlinx.coroutines.launch
 import kotlinx.parcelize.Parcelize
 import timber.log.Timber
 
@@ -56,7 +57,6 @@ class RoomFlowNode @AssistedInject constructor(
     private val roomDetailsEntryPoint: RoomDetailsEntryPoint,
     private val appNavigationStateService: AppNavigationStateService,
     roomMembershipObserver: RoomMembershipObserver,
-    coroutineScope: CoroutineScope,
 ) : BackstackNode<RoomFlowNode.NavTarget>(
     backstack = BackStack(
         initialElement = NavTarget.Messages,
@@ -76,9 +76,6 @@ class RoomFlowNode @AssistedInject constructor(
     ) : NodeInputs
 
     private val inputs: Inputs = inputs()
-    private val timeline = inputs.room.timeline()
-
-    private val roomFlowPresenter = RoomFlowPresenter(inputs.room)
 
     init {
         lifecycle.subscribe(
@@ -95,22 +92,34 @@ class RoomFlowNode @AssistedInject constructor(
             }
         )
 
+        lifecycleScope.fetchRoomMembers()
         roomMembershipObserver.updates
             .filter { update -> update.roomId == inputs.room.roomId && !update.isUserInRoom }
             .onEach {
                 navigateUp()
             }
-            .launchIn(coroutineScope)
+            .launchIn(lifecycleScope)
+    }
+
+    private fun CoroutineScope.fetchRoomMembers() = launch {
+        val room = inputs.room
+        room.fetchMembers()
+            .onFailure {
+                Timber.e(it, "Fail to fetch members for room ${room.roomId}")
+            }.onSuccess {
+                Timber.v("Success fetching members for room ${room.roomId}")
+            }
     }
 
     override fun resolve(navTarget: NavTarget, buildContext: BuildContext): Node {
         return when (navTarget) {
             NavTarget.Messages -> {
-                messagesEntryPoint.createNode(this, buildContext, object : MessagesEntryPoint.Callback {
+                val callback = object : MessagesEntryPoint.Callback {
                     override fun onRoomDetailsClicked() {
                         backstack.push(NavTarget.RoomDetails)
                     }
-                })
+                }
+                messagesEntryPoint.createNode(this, buildContext, callback)
             }
             NavTarget.RoomDetails -> {
                 roomDetailsEntryPoint.createNode(this, buildContext, emptyList())
@@ -128,7 +137,6 @@ class RoomFlowNode @AssistedInject constructor(
 
     @Composable
     override fun View(modifier: Modifier) {
-        roomFlowPresenter.present()
         Children(
             navModel = backstack,
             modifier = modifier,
