+++ features/roomdetails/impl/src/main/kotlin/io/element/android/features/roomdetails/impl/RoomDetailsNode.kt
@@ -41,7 +41,11 @@ class RoomDetailsNode @AssistedInject constructor(
     private val room: MatrixRoom,
 ) : Node(buildContext, plugins = plugins) {
 
-    private val callback = plugins<RoomDetailsFlowNode.Callback>().firstOrNull()
+    interface Callback : Plugin {
+        fun openRoomMemberList()
+    }
+
+    private val callback = plugins<Callback>().firstOrNull()
 
     private fun openRoomMemberList() {
         callback?.openRoomMemberList()
