+++ libraries/matrix/api/src/main/kotlin/io/element/android/libraries/matrix/api/room/MatrixRoom.kt
@@ -56,5 +56,5 @@ interface MatrixRoom: Closeable {
 
     suspend fun redactEvent(eventId: EventId, reason: String? = null): Result<Unit>
 
-    fun leave(): Result<Unit>
+    suspend fun leave(): Result<Unit>
 }
