+++ libraries/matrix/impl/src/main/kotlin/io/element/android/libraries/matrix/impl/room/RustMatrixRoom.kt
@@ -88,6 +88,7 @@ class RustMatrixRoom(
             coroutineDispatchers = coroutineDispatchers
         )
     }
+
     override fun close() {
         innerRoom.destroy()
         slidingSyncRoom.destroy()
@@ -183,7 +184,9 @@ class RustMatrixRoom(
         }
     }
 
-    override fun leave(): Result<Unit> {
-        return runCatching { innerRoom.leave() }
+    override suspend fun leave(): Result<Unit> = withContext(coroutineDispatchers.io) {
+        runCatching {
+            innerRoom.leave()
+        }
     }
 }
